scripts/requirements.txt
google-api-core==2.17.1
google-api-python-client==2.125.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-pubsub==2.22.0
google-cloud-secret-manager==2.18.2
googleapis-common-protos==1.62.0